refactor(MRUP): remove unneeded $() wrap; split into core and constant module
