+++ src/MarkRights-Userpage/modules/constant.ts
@@ -1,0 +1,12 @@
+export const SYSTEM_SCRIPT_LIST = [
+		'滥用过滤器',
+		'求闻编者',
+		'求闻百科社区',
+		'Example',
+		'External contributors',
+		'Maintenance script',
+		'MediaWiki default',
+		'MediaWiki message delivery',
+		'New user page',
+	],
+	WEBMASTER_LIST = ['QiuWen', 'QiuWen for Accounts', 'QiuwenSocialMedia'];
